requirements.txt
Django==1.11.6
djangorestframework==3.7.3
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pytz==2017.3
